+++ .integration/src/testnet1.rs
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
 
-use snarkvm_dpc::{testnet1::dpc::*, Account, AccountScheme, DPCScheme};
+use snarkvm_dpc::{testnet1::parameters::*, Account, AccountScheme, DPCScheme};
 
 use rand::{CryptoRng, Rng};
