+++ dpc/benches/noop_program.rs
@@ -19,7 +19,7 @@ extern crate criterion;
 
 use snarkvm_dpc::{
     prelude::*,
-    testnet2::{dpc::Testnet2Parameters, NoopProgram},
+    testnet2::{parameters::Testnet2Parameters, NoopProgram},
 };
 
 use criterion::Criterion;
