+++ dpc/src/testnet2/program/noop_program_circuit.rs
@@ -14,20 +14,20 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
 
-use crate::DPCComponents;
+use crate::Parameters;
 use snarkvm_algorithms::prelude::*;
 use snarkvm_gadgets::prelude::*;
 use snarkvm_r1cs::{Assignment, ConstraintSynthesizer, ConstraintSystem, SynthesisError};
 
 /// Always-accept program
-pub struct NoopCircuit<C: DPCComponents> {
+pub struct NoopCircuit<C: Parameters> {
     /// Commitment to the program input.
     pub local_data_root: Option<<C::LocalDataCRH as CRH>::Output>,
     /// Record position
     pub position: u8,
 }
 
-impl<C: DPCComponents> NoopCircuit<C> {
+impl<C: Parameters> NoopCircuit<C> {
     pub fn blank() -> Self {
         Self {
             local_data_root: Some(<C::LocalDataCRH as CRH>::Output::default()),
@@ -43,7 +43,7 @@ impl<C: DPCComponents> NoopCircuit<C> {
     }
 }
 
-impl<C: DPCComponents> ConstraintSynthesizer<C::InnerScalarField> for NoopCircuit<C> {
+impl<C: Parameters> ConstraintSynthesizer<C::InnerScalarField> for NoopCircuit<C> {
     fn generate_constraints<CS: ConstraintSystem<C::InnerScalarField>>(
         &self,
         cs: &mut CS,
