+++ parameters/examples/testnet2_universal_srs.rs
@@ -17,7 +17,7 @@
 use snarkvm_algorithms::crh::sha256::sha256;
 use snarkvm_dpc::{
     errors::DPCError,
-    testnet2::{dpc::Testnet2Parameters, ProgramSNARKUniversalSRS, Testnet2Components},
+    testnet2::{parameters::Testnet2Parameters, ProgramSNARKUniversalSRS, Testnet2Components},
 };
 use snarkvm_fields::ToConstraintField;
 use snarkvm_marlin::PolynomialCommitment;
